# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Holly API",
    "version": "1.0.0",
    "description": "API for Holly"
  },
  "paths": {
...
    "/_api/holly/git/clone": {
      "post": {
        "operationId": "holly_holly_api_views_git_clone_repository",
        "summary": "Clone Repository",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryResponse"
                }
              }
            }
          }
        },
        "description": "Clone a GitHub repository in the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Repository details including owner, name, and optional branch\n\nReturns:\n    GitRepositoryResponse: Result of the clone operation",
        "tags": ["Git"],
        "security": [{ "JWTAuth": [] }]
      }
    },
    "/_api/holly/git/worktree": {
      "post": {
        "operationId": "holly_holly_api_views_git_create_worktree",
        "summary": "Create Worktree",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryResponse"
                }
              }
            }
          }
        },
        "description": "Create a git worktree for a specific branch in the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Worktree details including repository and branch information\n\nReturns:\n    GitRepositoryResponse: Result of the worktree creation operation",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorktreeRequest" }
            }
          },
          "required": true
        },
        "security": [{ "JWTAuth": [] }]
      }
    },
    "/_api/holly/git/commit": {
      "post": {
        "operationId": "holly_holly_api_views_git_commit_changes",
        "summary": "Commit Changes",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryResponse"
                }
              }
            }
          }
        },
        "description": "Commit changes to a repository in the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Commit details including repository, branch, and message\n\nReturns:\n    GitRepositoryResponse: Result of the commit operation",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommitRequest" }
            }
          },
          "required": true
        },
        "security": [{ "JWTAuth": [] }]
      }
    },
    "/_api/holly/git/pull": {
      "post": {
        "operationId": "holly_holly_api_views_git_pull_changes",
        "summary": "Pull Changes",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryResponse"
                }
              }
            }
          }
        },
        "description": "Pull latest changes from remote repository in the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Pull details including repository and branch information\n\nReturns:\n    GitRepositoryResponse: Result of the pull operation",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PullRequest" }
            }
          },
          "required": true
        },
        "security": [{ "JWTAuth": [] }]
      }
    },
    "/_api/holly/git/push": {
      "post": {
        "operationId": "holly_holly_api_views_git_push_changes",
        "summary": "Push Changes",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryResponse"
                }
              }
            }
          }
        },
        "description": "Push local changes to remote repository from the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Push details including repository, branch, and force option\n\nReturns:\n    GitRepositoryResponse: Result of the push operation",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PushRequest" }
            }
          },
          "required": true
        },
        "security": [{ "JWTAuth": [] }]
      }
    },
    "/_api/holly/git/branches": {
      "post": {
        "operationId": "holly_holly_api_views_git_list_branches",
        "summary": "List Branches",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BranchesResponse" }
              }
            }
          }
        },
        "description": "Get a list of branches for a cloned repository in the mission container.\n\nArgs:\n    request: The HTTP request\n    mission_id: The UUID of the mission\n    repo_data: Repository details\n\nReturns:\n    BranchesResponse: The list of branches and other details",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryBranchesRequest"
              }
            }
          },
          "required": true
        },
        "security": [{ "JWTAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GitRepositoryResponse": {
        "description": "Schema for Git repository operation responses.",
        "properties": {
          "success": { "title": "Success", "type": "boolean" },
          "message": { "title": "Message", "type": "string" },
          "path": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Path"
          },
          "branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Branch"
          }
        },
        "required": ["success", "message"],
        "title": "GitRepositoryResponse",
        "type": "object"
      },
      "WorktreeRequest": {
        "description": "Request model for creating a worktree",
        "properties": {
          "repo_owner": { "title": "Repo Owner", "type": "string" },
          "repo_name": { "title": "Repo Name", "type": "string" },
          "auth_token": { "title": "Auth Token", "type": "string" },
          "branch": { "title": "Branch", "type": "string" },
          "create_branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Create Branch"
          },
          "base_branch": {
            "default": "main",
            "title": "Base Branch",
            "type": "string"
          }
        },
        "required": ["repo_owner", "repo_name", "auth_token", "branch"],
        "title": "WorktreeRequest",
        "type": "object"
      },
      "CommitRequest": {
        "description": "Request model for committing changes",
        "properties": {
          "repo_owner": { "title": "Repo Owner", "type": "string" },
          "repo_name": { "title": "Repo Name", "type": "string" },
          "auth_token": { "title": "Auth Token", "type": "string" },
          "branch": { "title": "Branch", "type": "string" },
          "create_branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Create Branch"
          },
          "commit_message": { "title": "Commit Message", "type": "string" },
          "files": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Files"
          }
        },
        "required": [
          "repo_owner",
          "repo_name",
          "auth_token",
          "branch",
          "commit_message"
        ],
        "title": "CommitRequest",
        "type": "object"
      },
      "PullRequest": {
        "description": "Request model for pulling changes",
        "properties": {
          "repo_owner": { "title": "Repo Owner", "type": "string" },
          "repo_name": { "title": "Repo Name", "type": "string" },
          "auth_token": { "title": "Auth Token", "type": "string" },
          "branch": { "title": "Branch", "type": "string" },
          "create_branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Create Branch"
          }
        },
        "required": ["repo_owner", "repo_name", "auth_token", "branch"],
        "title": "PullRequest",
        "type": "object"
      },
      "PushRequest": {
        "description": "Request model for pushing changes",
        "properties": {
          "repo_owner": { "title": "Repo Owner", "type": "string" },
          "repo_name": { "title": "Repo Name", "type": "string" },
          "auth_token": { "title": "Auth Token", "type": "string" },
          "branch": { "title": "Branch", "type": "string" },
          "create_branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Create Branch"
          },
          "force": { "default": false, "title": "Force", "type": "boolean" }
        },
        "required": ["repo_owner", "repo_name", "auth_token", "branch"],
        "title": "PushRequest",
        "type": "object"
      },
      "BranchesResponse": {
        "description": "Response model for repository branches",
        "properties": {
          "success": { "title": "Success", "type": "boolean" },
          "message": { "title": "Message", "type": "string" },
          "path": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Path"
          },
          "branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Branch"
          },
          "branches": {
            "items": { "type": "string" },
            "title": "Branches",
            "type": "array"
          },
          "current_branch": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Current Branch"
          }
        },
        "required": ["success", "message", "branches"],
        "title": "BranchesResponse",
        "type": "object"
      },
      "RepositoryBranchesRequest": {
        "description": "Request model for getting repository branches",
        "properties": {
          "repo_owner": { "title": "Repo Owner", "type": "string" },
          "repo_name": { "title": "Repo Name", "type": "string" }
        },
        "required": ["repo_owner", "repo_name"],
        "title": "RepositoryBranchesRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "JWTAuth": { "type": "http", "scheme": "bearer" },
      "AsyncJWTAuth": { "type": "http", "scheme": "bearer" }
    }
  },
  "servers": []
}